fix: Use a descriptive name like event & Prepare data in one place (App component) and pass it to child components.
